Move SaptuneDetails and SapSystemsOverview to pages
