+++ assets/js/pages/SapSystemsOverview/SapSystemsOverview.jsx
@@ -8,7 +8,7 @@ import { getEnsaVersionLabel } from '@lib/model/sapSystems';
 import PageHeader from '@common/PageHeader';
 import HealthIcon from '@pages/HealthIcon';
 import Table from '@common/Table';
-import SAPSystemItemOverview from '@components/SapSystemsOverview/SapSystemItemOverview';
+import SAPSystemItemOverview from '@pages/SapSystemsOverview/SapSystemItemOverview';
 import Tags from '@common/Tags';
 import HealthSummary from '@pages/HealthSummary';
 import DeregistrationModal from '@pages/DeregistrationModal';
