+++ assets/js/pages/SaptuneDetails/SaptuneDetailsPage.jsx
