+++ assets/js/pages/SaptuneDetails/SaptuneDetailsPage.test.jsx
