+++ assets/js/pages/SaptuneDetails/SaptuneVersion.jsx
