+++ assets/js/trento.jsx
@@ -17,13 +17,13 @@ import ClusterDetailsPage from '@pages/ClusterDetails';
 import ClusterSettingsPage from '@pages/ClusterSettingsPage';
 import { ExecutionResultsPage } from '@pages/ExecutionResults';
 import CheckResultDetailPage from '@pages/ExecutionResults/CheckResultDetail';
-import SapSystemsOverviewPage from '@components/SapSystemsOverview';
+import SapSystemsOverviewPage from '@pages/SapSystemsOverview';
 import HostDetailsPage, { HostSettingsPage } from '@pages/HostDetails';
 import DatabasesOverviewPage from '@pages/DatabasesOverview';
 import ChecksCatalogPage from '@pages/ChecksCatalog';
 import NotFound from '@pages/NotFound';
 import SomethingWentWrong from '@components/SomethingWentWrong';
-import SaptuneDetailsPage from '@components/SaptuneDetails';
+import SaptuneDetailsPage from '@pages/SaptuneDetails';
 import Settings from '@components/Settings';
 import Eula from '@pages/Eula';
 import Login from '@components/Login';
